integrated client and server

+++ expense-tracker-front/src/app/expense/components/expenses-overview/expenses-overview.component.ts
@@ -12,14 +12,16 @@ export class ExpensesOverviewComponent implements OnInit {
 
   expenses$: Observable<Expense[]> = of([]);
   
-  constructor(private expenseService: ExpenseService) { }
+  constructor(private expenseService: ExpenseService) { 
+    this.expenses$ = expenseService.getFindAllExpensesObservable();
+  }
 
   ngOnInit(): void {
-    this.expenses$ = this.expenseService.findAllExpenses();
+    this.expenseService.fetchAllExpenses();
   }
 
   deleteExpense(expense: Expense) {
-    if (confirm(`Are you sure you want to delete "${expense.name}"?`)) {
+    if (expense.id && confirm(`Are you sure you want to delete "${expense.name}"?`)) {
       this.expenseService.deleteExpense(expense.id);
     }
   }
